Move file tests from test_room_routes -> test_files
